refactor: split finder's concerns into focused modules
The finder still mixed the analysis pipeline with several unrelated
semantic concerns. Pull each into its own module, leaving Ciach as just
the orchestrator (discover → collect → check references → report).

New modules (all decoupled from Ciach, taking SourceIndex explicitly):
- reference_classifier.dart — ReferenceClassifier: the used/unused/doc-only
  decision, self-reference and pattern-match discounting
- remove_safety.dart       — RemoveSafety (was the in-file _RemoveSafety),
  now taking SourceIndex instead of the finder
- reference_kinds.dart      — isDocReference, a shared reference primitive
- conventions/freezed.dart  — FreezedUnions: annotation tracking + deser-only
  union-arm detection
- conventions/serialization.dart — isToJsonHook
- conventions/flutter_widgets.dart — State<Widget> pairing + coupled State
  removal
- paths.dart                — relativePosix, shared by finder and conventions

finder.dart drops from 832 to 485 lines; the run loop now reads as a
pipeline, with the suppression and removal-blocked decisions factored into
named helpers.

No behavioural change — all 84 tests pass, analyze and format clean.
