+++ lib/src/finder.dart
@@ -13,12 +13,17 @@ import 'dart:io';
 
 import 'package:ciach/src/candidates.dart';
 import 'package:ciach/src/concurrency.dart';
+import 'package:ciach/src/conventions/flutter_widgets.dart';
+import 'package:ciach/src/conventions/freezed.dart';
+import 'package:ciach/src/conventions/serialization.dart';
 import 'package:ciach/src/file_discovery.dart';
 import 'package:ciach/src/lsp/lsp_client.dart';
 import 'package:ciach/src/models.dart';
+import 'package:ciach/src/paths.dart';
+import 'package:ciach/src/reference_classifier.dart';
+import 'package:ciach/src/remove_safety.dart';
 import 'package:ciach/src/source_index.dart';
 import 'package:ciach/src/symbols.dart';
-import 'package:ciach/src/syntax_rules.dart';
 import 'package:path/path.dart' as p;
 import 'package:pro_lsp/pro_lsp.dart' show DocumentSymbol, Location;
 
@@ -28,6 +33,12 @@ import 'package:pro_lsp/pro_lsp.dart' show DocumentSymbol, Location;
 /// For every declaration reported by `textDocument/documentSymbol`, a
 /// `textDocument/references` query is issued at the declaration's name, with
 /// `includeDeclaration: false`. An empty result means the declaration is unused.
+///
+/// `Ciach` owns the pipeline (discover → collect → check references → report);
+/// the semantic pieces live in collaborators: [ReferenceClassifier] decides
+/// used/unused, [RemoveSafety] flags findings that can't be auto-removed, and
+/// the `conventions/` rules ([FreezedUnions], serialization and Flutter
+/// widgets) keep framework-driven declarations alive.
 class Ciach {
   /// Creates a finder that runs with the given [options].
   Ciach(this.options);
@@ -39,22 +50,13 @@ class Ciach {
   /// shared by the classification and the structural detectors.
   final _sources = SourceIndex();
 
-  /// Keys `(path, class name)` of classes annotated `@freezed`/`@Freezed`.
-  final _freezedAnnotatedClasses = <DeclKey>{};
+  /// Freezed-union tracking, fed as candidates are collected.
+  final _freezed = FreezedUnions();
 
-  static final _freezedAnnotation = RegExp(r'@(?:freezed|Freezed)\b');
-
-  /// A JSON-value return type on a `toJson`'s declaration line, before the name.
-  static final _toJsonJsonReturn = RegExp(
-    r'\b(?:Map|List|String|int|double|num|bool|Object|dynamic)\b\s*(?:<[^;{]*>)?\s*\??\s*$',
+  late final _classifier = ReferenceClassifier(
+    _sources,
+    unusedUnionMembers: options.unusedUnionMembers,
   );
-
-  /// The `State<` immediately preceding a type argument, with a token boundary
-  /// before `State` so `MyState<…>`/`FooState<…>` don't match.
-  static final _statePrefix = RegExp(r'(?:^|[^A-Za-z0-9_$])State<\s*$');
-
-  /// The `>` that closes a single `State<…>` type argument.
-  static final _stateSuffix = RegExp(r'^\s*>');
 
   /// Advisory note attached to a sole, zero-parameter private constructor
   /// (`Foo._();`) — the classic prevent-instantiation marker. Such a
@@ -130,43 +132,24 @@ class Ciach {
       // and a line is emitted as soon as its last declaration is checked. Files
       // with no candidates are already counted as done.
       _report('Checking references for $declarationsChecked declaration(s)…');
-      final remainingPerFile = <String, int>{};
-      for (final candidate in candidates) {
-        remainingPerFile.update(
-          candidate.path,
-          (n) => n + 1,
-          ifAbsent: () => 1,
-        );
-      }
-      final totalFiles = files.length;
-      var filesDone = totalFiles - remainingPerFile.length;
-
-      final refsByCandidate = await mapPooled(candidates, options.concurrency, (
-        candidate,
-      ) async {
-        final refs = await client.references(
-          candidate.uri,
-          candidate.symbol.selectionRange.start,
-        );
-        if (remainingPerFile.update(candidate.path, (n) => n - 1) == 0) {
-          filesDone++;
-          _report(
-            '[$filesDone/$totalFiles] '
-            '${p.relative(candidate.path, from: rootPath)}',
-          );
-        }
-        return refs;
-      });
+      final refsByCandidate = await _checkReferences(
+        client,
+        candidates,
+        files.length,
+        rootPath,
+      );
 
       final statuses = [
         for (var i = 0; i < candidates.length; i++)
-          _classify(candidates[i], refsByCandidate[i]),
+          _classifier.classify(candidates[i], refsByCandidate[i]),
       ];
 
-      // A deser-only union arm reads zero references but is a live serialization member.
-      final freezedUnionArms = _freezedDeserializedUnionArms(
+      // A deser-only union arm reads zero references but is a live serialization
+      // member.
+      final freezedUnionArms = _freezed.deserializationOnlyArms(
         candidates,
         statuses,
+        _sources,
       );
 
       // Names of classes flagged unused, per file. A whole dead class is
@@ -182,8 +165,8 @@ class Ciach {
         }
       }
 
-      final safety = _RemoveSafety.analyze(
-        this,
+      final safety = RemoveSafety.analyze(
+        _sources,
         candidates,
         statuses,
         refsByCandidate,
@@ -195,63 +178,30 @@ class Ciach {
         final refs = refsByCandidate[i];
         switch (statuses[i]) {
           case .unused:
-            if (freezedUnionArms.contains(i)) {
-              break;
-            }
-            if (!options.reportToJson && _isToJsonHook(candidate)) {
-              break;
-            }
-            if (_isConstructorOfDeadClass(candidate, deadClassNames)) {
-              break;
-            }
-            final containerKey = candidate.containerKey;
-            // An enum value reached only through `.values` iteration is used,
-            // not dead: suppress it entirely (never reported) rather than
-            // flagging it for the empty-enum guard.
-            if (candidate.isEnumValue &&
-                containerKey != null &&
-                safety.enumValuesIterated.contains(containerKey)) {
+            if (_isSuppressed(
+              candidate,
+              i,
+              freezedUnionArms,
+              deadClassNames,
+              safety,
+            )) {
               break;
             }
             final isClass = candidate.symbol.kind == .class$;
-            final paired = isClass
-                ? _pairedStateRemovals(
-                    candidate,
-                    refs,
-                    candidates,
-                    refsByCandidate,
-                    rootPath,
-                  )
-                : const <CoupledRemoval>[];
-            // A finding is report-only (removalBlocked) when auto-removing it
-            // would break the build:
-            //
-            // * a class kept dead only by type patterns under
-            //   --unused-union-members (never constructed, only matched):
-            //   deleting a sealed member and its scattered `case`s is a source
-            //   rewrite this tool won't attempt;
-            // * an enum value whose removal would empty a still-referenced enum;
-            // * the last constructor of a live class with `final` fields or
-            //   super-constructor forwarding.
-            //
-            // Each is surfaced so a human can act on it, but the remover leaves
-            // it — and anything coupled to it — entirely alone.
-            final removalBlocked =
-                (isClass &&
-                    options.unusedUnionMembers &&
-                    _isPatternMatchedClass(candidate, refs)) ||
-                (candidate.isEnumValue &&
-                    containerKey != null &&
-                    safety.emptiedEnums.contains(containerKey)) ||
-                (candidate.symbol.kind == .constructor &&
-                    containerKey != null &&
-                    safety.blockedCtorClasses.contains(containerKey));
             unused.add(
               _toUnused(
                 candidate,
                 rootPath,
-                coupledRemovals: paired,
-                removalBlocked: removalBlocked,
+                coupledRemovals: isClass
+                    ? _sources.pairedStateRemovals(
+                        candidate,
+                        refs,
+                        candidates,
+                        refsByCandidate,
+                        rootPath,
+                      )
+                    : const [],
+                removalBlocked: _isRemovalBlocked(candidate, refs, safety),
                 // A sole, zero-parameter private constructor is dead code like
                 // any other, but nudge toward `abstract final class`.
                 hint: candidate.isPreventInstantiationCtor
@@ -279,6 +229,91 @@ class Ciach {
       declarationsChecked: declarationsChecked,
       elapsed: stopwatch.elapsed,
     );
+  }
+
+  /// Queries `textDocument/references` for every candidate through one global
+  /// pool, reporting `[done/total]` progress as each file's last query lands.
+  Future<List<List<Location>>> _checkReferences(
+    LspClient client,
+    List<Candidate> candidates,
+    int totalFiles,
+    String rootPath,
+  ) {
+    final remainingPerFile = <String, int>{};
+    for (final candidate in candidates) {
+      remainingPerFile.update(candidate.path, (n) => n + 1, ifAbsent: () => 1);
+    }
+    var filesDone = totalFiles - remainingPerFile.length;
+
+    return mapPooled(candidates, options.concurrency, (candidate) async {
+      final refs = await client.references(
+        candidate.uri,
+        candidate.symbol.selectionRange.start,
+      );
+      if (remainingPerFile.update(candidate.path, (n) => n - 1) == 0) {
+        filesDone++;
+        _report(
+          '[$filesDone/$totalFiles] '
+          '${p.relative(candidate.path, from: rootPath)}',
+        );
+      }
+      return refs;
+    });
+  }
+
+  /// Whether an unused [candidate] should be silently suppressed (never
+  /// reported): a live freezed-union arm, an exempt `toJson` hook, a
+  /// constructor removed with its already-dead class, or an enum value reached
+  /// only through `.values` iteration.
+  bool _isSuppressed(
+    Candidate candidate,
+    int index,
+    Set<int> freezedUnionArms,
+    Map<String, Set<String>> deadClassNames,
+    RemoveSafety safety,
+  ) {
+    if (freezedUnionArms.contains(index)) {
+      return true;
+    }
+    if (!options.reportToJson && _sources.isToJsonHook(candidate)) {
+      return true;
+    }
+    if (_isConstructorOfDeadClass(candidate, deadClassNames)) {
+      return true;
+    }
+    final containerKey = candidate.containerKey;
+    return candidate.isEnumValue &&
+        containerKey != null &&
+        safety.enumValuesIterated.contains(containerKey);
+  }
+
+  /// Whether a dead [candidate] is real but must *not* be auto-removed, because
+  /// doing so would break the build:
+  ///
+  /// * a class kept dead only by type patterns under `--unused-union-members`
+  ///   (never constructed, only matched): deleting a sealed member and its
+  ///   scattered `case`s is a source rewrite this tool won't attempt;
+  /// * an enum value whose removal would empty a still-referenced enum;
+  /// * the last constructor of a live class with `final` fields or
+  ///   super-constructor forwarding.
+  ///
+  /// Each is surfaced so a human can act on it, but the remover leaves it — and
+  /// anything coupled to it — entirely alone.
+  bool _isRemovalBlocked(
+    Candidate candidate,
+    List<Location> refs,
+    RemoveSafety safety,
+  ) {
+    final containerKey = candidate.containerKey;
+    return (candidate.symbol.kind == .class$ &&
+            options.unusedUnionMembers &&
+            _classifier.isPatternMatchedClass(candidate, refs)) ||
+        (candidate.isEnumValue &&
+            containerKey != null &&
+            safety.emptiedEnums.contains(containerKey)) ||
+        (candidate.symbol.kind == .constructor &&
+            containerKey != null &&
+            safety.blockedCtorClasses.contains(containerKey));
   }
 
   /// Opens [path] in the analysis server without collecting candidates from
@@ -335,10 +370,7 @@ class Ciach {
     List<Candidate> out,
   ) {
     for (final symbol in symbols) {
-      if (typeLikeKinds.contains(symbol.kind) &&
-          _freezedAnnotation.hasMatch(symbol.leadingMetadata(lines))) {
-        _freezedAnnotatedClasses.add(DeclKey(path, symbol.name));
-      }
+      _freezed.noteIfAnnotated(path, symbol, lines);
       if (_shouldConsider(symbol, parentIsEnum, lines)) {
         out.add(
           Candidate(
@@ -421,7 +453,7 @@ class Ciach {
     return .new(
       name: name,
       kind: symbol.reportedKind(parentIsEnum: candidate.isEnumValue),
-      filePath: _relPath(candidate.path, rootPath),
+      filePath: relativePosix(candidate.path, rootPath),
       // LSP positions are zero-based; report them one-based for humans.
       line: start.line + 1,
       column: start.character + 1,
@@ -435,277 +467,12 @@ class Ciach {
     );
   }
 
-  /// Classifies a candidate from the references reported for it.
-  ///
-  /// Non-class candidates keep the simple rule: any real (non-doc) reference
-  /// means used, only doc-comment links means doc-only, none means unused.
-  ///
-  /// Classes get [_classifyClass], which discounts *self-references* — the
-  /// class's own body, and the `State<Self>` StatefulWidget pairing — so a
-  /// class kept alive only by its own unnamed constructor's declaration (whose
-  /// name coincides with the class) is correctly seen as dead.
-  RefStatus _classify(Candidate candidate, List<Location> refs) {
-    if (candidate.symbol.kind == .class$) {
-      return _classifyClass(candidate, refs);
-    }
-    if (refs.isEmpty) {
-      return .unused;
-    }
-    return refs.any((loc) => !_isDocReference(loc)) ? .used : .docOnly;
-  }
-
-  /// Classifies a class by its references, ignoring self-references.
-  ///
-  /// A class is used only if some reference is a real (non-doc) reference from
-  /// *outside* the class itself; if the only outside references are doc-comment
-  /// links it is doc-only, and otherwise (only self-references, or none at all)
-  /// it is unused. This is deliberately conservative: any single unexplained
-  /// outside reference keeps the class alive, so the failure mode is missing a
-  /// dead class, never deleting a live one.
-  RefStatus _classifyClass(Candidate candidate, List<Location> refs) {
-    var hasExternalDoc = false;
-    var hasPatternMatch = false;
-    for (final loc in refs) {
-      if (_isSelfClassReference(candidate, loc)) {
-        continue;
-      }
-      if (_isDocReference(loc)) {
-        hasExternalDoc = true;
-        continue;
-      }
-      // With --unused-union-members, a reference that is confidently a *type
-      // pattern* (a `case`/if-case/while-case pattern, or a switch-expression
-      // arm) is a *match*, not a construction: if the type is never
-      // constructed, no such match can ever fire, so it is discounted like a
-      // self-reference. Any reference that is not confidently a type pattern
-      // keeps the class alive — the conservative choice (nested sub-patterns
-      // and pattern-variable declarations are intentionally not recognized).
-      if (options.unusedUnionMembers && _sources.isPatternRef(loc)) {
-        hasPatternMatch = true;
-        continue;
-      }
-      // A real, non-pattern reference from outside: the class is used.
-      return .used;
-    }
-    // Matched-only-by-a-pattern (never constructed) is dead code, not a softer
-    // doc-only report.
-    if (hasPatternMatch) {
-      return .unused;
-    }
-    return hasExternalDoc ? .docOnly : .unused;
-  }
-
-  /// Whether [location] points at a dartdoc `[Xxx]`-style reference rather than
-  /// real code — i.e. its line, in the file it points into, starts with `///`.
-  /// Block (`/** */`) doc comments aren't recognized; `///` is the standard and
-  /// lint-enforced style.
-  bool _isDocReference(Location location) {
-    final lines = _sources.lines(SourceIndex.pathOf(location.uri));
-    final line = location.range.start.line;
-    if (line < 0 || line >= lines.length) {
-      return false;
-    }
-    return lines[line].trim().startsWith('///');
-  }
-
-  /// Whether [loc] is a reference to [candidate] that does not count as a use.
-  ///
-  /// Two shapes qualify:
-  ///
-  /// 1. A reference inside the class's own source span — its body, signature,
-  ///    or leading doc/annotation lines. This covers the unnamed constructor's
-  ///    declaration, a `State<Foo>` return type on the widget's own
-  ///    `createState`, and any purely-internal self-use.
-  /// 2. A `State<Foo>` type-argument reference anywhere — the StatefulWidget
-  ///    pairing. `State<Foo>` can only ever denote the state object of the
-  ///    `Foo` widget, so it never means `Foo` itself is used elsewhere.
-  bool _isSelfClassReference(Candidate candidate, Location loc) {
-    if (SourceIndex.pathOf(loc.uri) == candidate.path) {
-      final top = candidate.symbol.metadataTopLine(
-        _sources.lines(candidate.path),
-      );
-      final pos = loc.range.start;
-      if (pos.line >= top && pos.atOrBefore(candidate.symbol.range.end)) {
-        return true;
-      }
-    }
-    return _isStatePairingReference(candidate.symbol.name, loc);
-  }
-
-  /// Whether [loc] is the class name [className] appearing as the sole type
-  /// argument of `State<…>`, e.g. `class _FooState extends State<Foo>`.
-  bool _isStatePairingReference(String className, Location loc) {
-    final start = loc.range.start;
-    final end = loc.range.end;
-    if (start.line != end.line) {
-      return false;
-    }
-    final lines = _sources.lines(SourceIndex.pathOf(loc.uri));
-    if (start.line < 0 || start.line >= lines.length) {
-      return false;
-    }
-    final line = lines[start.line];
-    if (start.character < 0 ||
-        end.character > line.length ||
-        start.character > end.character) {
-      return false;
-    }
-    if (line.substring(start.character, end.character) != className) {
-      return false;
-    }
-    return _statePrefix.hasMatch(line.substring(0, start.character)) &&
-        _stateSuffix.hasMatch(line.substring(end.character));
-  }
-
   bool _isConstructorOfDeadClass(
     Candidate candidate,
     Map<String, Set<String>> deadClassNames,
   ) =>
       candidate.symbol.kind == .constructor &&
       (deadClassNames[candidate.path]?.contains(candidate.container) ?? false);
-
-  /// Whether [candidate] is a `toJson()` serialization hook — a zero-required-arg
-  /// method named `toJson` returning any JSON value (`Map`/`List`/`String`/`num`/
-  /// `int`/`double`/`bool`, or `Object`/`dynamic`). `jsonEncode(obj)` dispatches to
-  /// it dynamically with no source-level `.toJson()` token, so the reference search
-  /// can't see that use; exempt it for any class, annotated or not.
-  bool _isToJsonHook(Candidate candidate) {
-    final symbol = candidate.symbol;
-    if (symbol.kind != .method ||
-        symbol.name != 'toJson' ||
-        !symbol.hasNoParameters) {
-      return false;
-    }
-    final lines = _sources.lines(candidate.path);
-    final line = symbol.selectionRange.start.line;
-    if (line < 0 || line >= lines.length) {
-      return false;
-    }
-    final col = symbol.selectionRange.start.character;
-    final text = lines[line];
-    final beforeName = col <= text.length ? text.substring(0, col) : text;
-    return _toJsonJsonReturn.hasMatch(beforeName);
-  }
-
-  /// Redirecting-factory arms of a `@freezed`/`@Freezed` union with a referenced
-  /// `fromJson`; conservative — never-dispatched arms are kept too, being
-  /// statically indistinguishable.
-  Set<int> _freezedDeserializedUnionArms(
-    List<Candidate> candidates,
-    List<RefStatus> statuses,
-  ) {
-    final unionsWithUsedFromJson = <DeclKey>{};
-    for (var i = 0; i < candidates.length; i++) {
-      final c = candidates[i];
-      if (c.symbol.kind == .constructor &&
-          statuses[i] == .used &&
-          c.symbol.declarationName(c.container) == 'fromJson') {
-        if (c.containerKey case final key?) {
-          unionsWithUsedFromJson.add(key);
-        }
-      }
-    }
-
-    final arms = <int>{};
-    for (var i = 0; i < candidates.length; i++) {
-      if (statuses[i] != .unused) {
-        continue;
-      }
-      final c = candidates[i];
-      if (c.symbol.kind != .constructor) {
-        continue;
-      }
-      if (c.containerKey case final key?
-          when _freezedAnnotatedClasses.contains(key) &&
-              unionsWithUsedFromJson.contains(key) &&
-              _sources.isRedirectingFactory(c)) {
-        arms.add(i);
-      }
-    }
-    return arms;
-  }
-
-  /// The extra spans to remove alongside a dead [widget] class: the paired
-  /// private `State<Widget>` subclass, when there is exactly one and it is used
-  /// only from within the widget (via `createState`). Returns an empty list for
-  /// a plain class, or a StatefulWidget whose State is referenced elsewhere.
-  ///
-  /// Removing the widget on its own would leave
-  /// `class _S extends State<Widget>` referring to a now-deleted type — a build
-  /// break — so the State is coupled to the widget's removal, but it is not
-  /// itself reported.
-  List<CoupledRemoval> _pairedStateRemovals(
-    Candidate widget,
-    List<Location> widgetRefs,
-    List<Candidate> candidates,
-    List<List<Location>> refsByCandidate,
-    String rootPath,
-  ) {
-    final out = <CoupledRemoval>[];
-    for (final loc in widgetRefs) {
-      if (!_isStatePairingReference(widget.symbol.name, loc) ||
-          SourceIndex.pathOf(loc.uri) != widget.path) {
-        continue;
-      }
-      // The widget's own `createState` return type is inside the widget and
-      // removed with it; only a pairing reference outside the widget points at
-      // the separate State subclass.
-      if (loc.range.start.within(widget.symbol)) {
-        continue;
-      }
-      for (var j = 0; j < candidates.length; j++) {
-        final state = candidates[j];
-        if (state.symbol.kind != .class$ ||
-            state.path != widget.path ||
-            identical(state, widget) ||
-            !loc.range.start.within(state.symbol)) {
-          continue;
-        }
-        if (_referencedOnlyWithin(
-          refsByCandidate[j],
-          widget.symbol,
-          widget.path,
-        )) {
-          out.add((
-            filePath: _relPath(state.path, rootPath),
-            range: state.symbol.declarationRange,
-          ));
-        }
-        break;
-      }
-    }
-    return out;
-  }
-
-  /// Whether every *code* reference in [refs] lies within [enclosing] in
-  /// [path] — used to confirm a paired State subclass is reachable only from
-  /// its widget. Doc-comment links are ignored: documentation never keeps code
-  /// alive, so it must not block coupling.
-  bool _referencedOnlyWithin(
-    List<Location> refs,
-    DocumentSymbol enclosing,
-    String path,
-  ) => refs.every(
-    (loc) =>
-        _isDocReference(loc) ||
-        (SourceIndex.pathOf(loc.uri) == path &&
-            loc.range.start.within(enclosing)),
-  );
-
-  /// Whether [candidate] — already classified as unused under
-  /// `--unused-union-members` — is kept dead by *type patterns*: at least one
-  /// of its non-self, non-doc references is a `case`/switch-expression pattern
-  /// match rather than a construction.
-  bool _isPatternMatchedClass(Candidate candidate, List<Location> refs) =>
-      refs.any(
-        (loc) =>
-            !_isSelfClassReference(candidate, loc) &&
-            !_isDocReference(loc) &&
-            _sources.isPatternRef(loc),
-      );
-
-  String _relPath(String absPath, String rootPath) =>
-      p.split(p.relative(absPath, from: rootPath)).join('/');
 
   static int _byLocation(UnusedDeclaration a, UnusedDeclaration b) {
     final byFile = a.filePath.compareTo(b.filePath);
@@ -715,117 +482,4 @@ class Ciach {
     final byLine = a.line.compareTo(b.line);
     return byLine != 0 ? byLine : a.column.compareTo(b.column);
   }
-}
-
-/// The remove-safety pre-pass: facts gathered up front so the reporting loop
-/// stays a set of cheap lookups when deciding which findings are report-only
-/// (`removalBlocked`) because auto-removing them would break the build.
-///
-/// * [emptiedEnums] — enums every one of whose values would be removed while
-///   the enum type is still referenced, leaving `enum E {}` (a compile error).
-/// * [blockedCtorClasses] — live classes all of whose constructors are dead:
-///   removing them synthesizes an implicit default constructor that strands
-///   `final` fields or breaks super-constructor forwarding.
-/// * [enumValuesIterated] — enums whose values are all reachable through
-///   `.values` iteration, so a value reached only that way is used, not dead,
-///   and is suppressed entirely rather than reported.
-class _RemoveSafety {
-  const _RemoveSafety({
-    required this.emptiedEnums,
-    required this.blockedCtorClasses,
-    required this.enumValuesIterated,
-  });
-
-  factory _RemoveSafety.analyze(
-    Ciach finder,
-    List<Candidate> candidates,
-    List<RefStatus> statuses,
-    List<List<Location>> refsByCandidate,
-    Map<String, Set<String>> deadClassNames,
-  ) {
-    final sources = finder._sources;
-    final enumTypeHasRef = <DeclKey, bool>{};
-    final enumValueTotal = <DeclKey, int>{};
-    final enumValueDead = <DeclKey, int>{};
-    final enumValuesIterated = <DeclKey>{};
-    final ctorTotal = <DeclKey, int>{};
-    final ctorDead = <DeclKey, int>{};
-    final ctorForwardsSuper = <DeclKey, bool>{};
-    final classByKey = <DeclKey, Candidate>{};
-
-    for (var i = 0; i < candidates.length; i++) {
-      final candidate = candidates[i];
-      final symbol = candidate.symbol;
-      final unused = statuses[i] == .unused;
-      if (symbol.kind == .enum$ && !candidate.isEnumValue) {
-        enumTypeHasRef[candidate.key] = refsByCandidate[i].isNotEmpty;
-        if (refsByCandidate[i].any(sources.isDotValuesRef) ||
-            sources.enumIteratesOwnValues(candidate)) {
-          enumValuesIterated.add(candidate.key);
-        }
-      } else if (candidate.isEnumValue) {
-        if (candidate.containerKey case final key?) {
-          enumValueTotal.update(key, (n) => n + 1, ifAbsent: () => 1);
-          if (unused) {
-            enumValueDead.update(key, (n) => n + 1, ifAbsent: () => 1);
-          }
-        }
-      } else if (symbol.kind == .class$) {
-        classByKey[candidate.key] = candidate;
-      } else if (symbol.kind == .constructor) {
-        if (candidate.containerKey case final key?) {
-          ctorTotal.update(key, (n) => n + 1, ifAbsent: () => 1);
-          if (unused) {
-            ctorDead.update(key, (n) => n + 1, ifAbsent: () => 1);
-            if (sources.ctorForwardsSuper(candidate)) {
-              ctorForwardsSuper[key] = true;
-            }
-          }
-        }
-      }
-    }
-
-    final emptiedEnums = <DeclKey>{};
-    for (final MapEntry(:key, value: total) in enumValueTotal.entries) {
-      final dead = enumValueDead[key] ?? 0;
-      if (dead == 0 || dead != total) {
-        continue;
-      }
-      // Conservative: if the enum-type candidate is missing we cannot prove the
-      // enum is itself being removed, so assume it stays and block.
-      if (enumTypeHasRef[key] ?? true) {
-        emptiedEnums.add(key);
-      }
-    }
-
-    final blockedCtorClasses = <DeclKey>{};
-    for (final MapEntry(:key, value: total) in ctorTotal.entries) {
-      final dead = ctorDead[key] ?? 0;
-      if (dead == 0 || dead != total) {
-        continue;
-      }
-      // A dead class is removed whole (its constructors go with it), so its
-      // constructors are never reported on their own — nothing to guard.
-      if (deadClassNames[key.path]?.contains(key.name) ?? false) {
-        continue;
-      }
-      final classCandidate = classByKey[key];
-      final hasFinalField =
-          classCandidate != null &&
-          sources.classHasFinalInstanceField(classCandidate);
-      if (hasFinalField || (ctorForwardsSuper[key] ?? false)) {
-        blockedCtorClasses.add(key);
-      }
-    }
-
-    return _RemoveSafety(
-      emptiedEnums: emptiedEnums,
-      blockedCtorClasses: blockedCtorClasses,
-      enumValuesIterated: enumValuesIterated,
-    );
-  }
-
-  final Set<DeclKey> emptiedEnums;
-  final Set<DeclKey> blockedCtorClasses;
-  final Set<DeclKey> enumValuesIterated;
 }
